Add files via upload
Farmers in our country suffer a huge disadvange due to the lack of knowledge of science and technology which would definitely help them better their yield and improvise their farming practices. This is a contribution from me and my three mates - Himani Popli,Rishika Gupta and Lubhani Agarwal - as an effort/a proposal to be of some use to the feeders of the nation - our farmers. Interested audience are requested to unzip the files and check out our solutions as learning web developers. 
Feedbacks and recommended changes are heartily welcomed